// Capstone Project/src/app/pages/admin/admin-posting-cost-management-page/screen/types.ts
export enum adminPagePostingCostEnum {
  GET_DATA_POSTING_COST_REQUEST = 'APP/ADMIN_PAGE/GET_DATA_POSTING_COST_REQUEST',
  GET_DATA_POSTING_COST_SUCCESS = 'APP/ADMIN_PAGE/GET_DATA_POSTING_COST_SUCCESS',
  GET_DATA_POSTING_COST_FAILURE = 'APP/ADMIN_PAGE/GET_DATA_POSTING_COST_FAILURE',

  CREATE_DATA_POSTING_COST_REQUEST = 'APP/ADMIN_PAGE/CREATE_DATA_POSTING_COST_REQUEST',
  CREATE_DATA_POSTING_COST_SUCCESS = 'APP/ADMIN_PAGE/CREATE_DATA_POSTING_COST_SUCCESS',
  CREATE_DATA_POSTING_COST_FAILURE = 'APP/ADMIN_PAGE/CREATE_DATA_POSTING_COST_FAILURE',

  UPDATE_DATA_POSTING_COST_REQUEST = 'APP/ADMIN_PAGE/UPDATE_DATA_POSTING_COST_REQUEST',
  UPDATE_DATA_POSTING_COST_SUCCESS = 'APP/ADMIN_PAGE/UPDATE_DATA_POSTING_COST_SUCCESS',
  UPDATE_DATA_POSTING_COST_FAILURE = 'APP/ADMIN_PAGE/UPDATE_DATA_POSTING_COST_FAILURE',

  DELETE_DATA_POSTING_COST_REQUEST = 'APP/ADMIN_PAGE/DELETE_DATA_POSTING_COST_REQUEST',
  DELETE_DATA_POSTING_COST_SUCCESS = 'APP/ADMIN_PAGE/DELETE_DATA_POSTING_COST_SUCCESS',
  DELETE_DATA_POSTING_COST_FAILURE = 'APP/ADMIN_PAGE/DELETE_DATA_POSTING_COST_FAILURE',

  CLEAR_POSTING_COST_STATE = 'APP/ADMIN_PAGE/CLEAR_POSTING_COST_STATE',
}